llm_integration/requirements.txt
faiss-cpu
openai
python-dotenv
numpy
streamlit